refactored parkour pathway
